removes circular dep
