+++ tests/models/test_privacy_request.py
@@ -1,7 +1,5 @@
 from datetime import datetime, timedelta
 from typing import List
-from unittest import mock
-from unittest.mock import Mock
 from uuid import uuid4
 
 from sqlalchemy.orm import Session
@@ -9,7 +7,6 @@ from sqlalchemy.orm import Session
 from fidesops.models.policy import Policy
 from fidesops.models.privacy_request import (
     PrivacyRequest,
-    PrivacyRequestRunner,
     PrivacyRequestStatus,
 )
 from fidesops.schemas.redis_cache import PrivacyRequestIdentity
@@ -132,36 +129,3 @@ def test_delete_privacy_request_removes_cached_data(
     from_db = PrivacyRequest.get(db=db, id=privacy_request.id)
     assert from_db is None
     assert cache.get(key) is None
-
-
-@mock.patch("fidesops.task.graph_task.run_access_request")
-@mock.patch("fidesops.models.privacy_request.upload")
-def test_policy_upload_called(
-    upload_mock: Mock,
-    run_access_request_mock: Mock,
-    db: Session,
-    privacy_request: PrivacyRequest,
-    privacy_request_runner: PrivacyRequestRunner,
-) -> None:
-    privacy_request_runner.run()
-    assert privacy_request.finished_processing_at is not None
-    assert upload_mock.called
-    assert run_access_request_mock.called
-
-
-def test_start_processing_sets_started_processing_at(
-    db: Session,
-    privacy_request: PrivacyRequest,
-) -> None:
-    privacy_request.started_processing_at = None
-    privacy_request.start_processing(db=db)
-    assert privacy_request.started_processing_at is not None
-
-
-def test_start_processing_doesnt_overwrite_started_processing_at(
-    db: Session,
-    privacy_request: PrivacyRequest,
-) -> None:
-    before = privacy_request.started_processing_at
-    privacy_request.start_processing(db=db)
-    assert privacy_request.started_processing_at == before
