+++ tests/service/privacy_request/onetrust_test.py
@@ -21,19 +21,18 @@ from fidesops.models.privacy_request import PrivacyRequest
 from fidesops.models.storage import StorageConfig
 from fidesops.schemas.third_party.onetrust import OneTrustRequest, OneTrustSubtask
 from fidesops.schemas.storage.storage import StorageSecrets, StorageDetails
-from fidesops.service.request_intake.onetrust_service import (
+from fidesops.service.privacy_request.onetrust_service import (
     ONETRUST_POLICY_KEY,
-    FIDES_TASK,
-    intake_onetrust_requests,
+    FIDES_TASK, OneTrustService,
 )
 
 
 @mock.patch(
-    "fidesops.service.request_intake.onetrust_service.get_onetrust_access_token"
+    "fidesops.service.privacy_request.onetrust_service.get_onetrust_access_token"
 )
-@mock.patch("fidesops.service.request_intake.onetrust_service._get_all_requests")
-@mock.patch("fidesops.service.request_intake.onetrust_service._get_all_subtasks")
-@mock.patch("fidesops.models.privacy_request.PrivacyRequestRunner.run")
+@mock.patch("fidesops.service.privacy_request.onetrust_service.OneTrustService._get_all_requests")
+@mock.patch("fidesops.service.privacy_request.onetrust_service.OneTrustService._get_all_subtasks")
+@mock.patch("fidesops.service.privacy_request.request_runner_service.PrivacyRequestRunner.run")
 def test_intake_onetrust_requests_success(
     finish_processing_mock: Mock,
     mock_get_all_subtasks: Mock,
@@ -70,7 +69,7 @@ def test_intake_onetrust_requests_success(
     mock_get_all_subtasks.side_effect = [[mock_subtask_1], [mock_subtask_2]]
 
     # call test function
-    intake_onetrust_requests(storage_config_onetrust.key)
+    OneTrustService.intake_onetrust_requests(storage_config_onetrust.key)
 
     # assert
     mock_get_all_requests.assert_called_with(
@@ -109,11 +108,11 @@ def test_intake_onetrust_requests_success(
 
 
 @mock.patch(
-    "fidesops.service.request_intake.onetrust_service.get_onetrust_access_token"
+    "fidesops.service.privacy_request.onetrust_service.get_onetrust_access_token"
 )
-@mock.patch("fidesops.service.request_intake.onetrust_service._get_all_requests")
-@mock.patch("fidesops.service.request_intake.onetrust_service._get_all_subtasks")
-@mock.patch("fidesops.models.privacy_request.PrivacyRequestRunner.run")
+@mock.patch("fidesops.service.privacy_request.onetrust_service.OneTrustService._get_all_requests")
+@mock.patch("fidesops.service.privacy_request.onetrust_service.OneTrustService._get_all_subtasks")
+@mock.patch("fidesops.service.privacy_request.request_runner_service.PrivacyRequestRunner.run")
 def test_intake_onetrust_requests_no_config(
     finish_processing_mock: Mock,
     mock_get_all_subtasks: Mock,
@@ -148,7 +147,7 @@ def test_intake_onetrust_requests_no_config(
     mock_get_all_subtasks.side_effect = [[mock_subtask_1], [mock_subtask_2]]
 
     with pytest.raises(StorageConfigNotFoundException):
-        intake_onetrust_requests(storage_config_onetrust.key)
+        OneTrustService.intake_onetrust_requests(storage_config_onetrust.key)
 
     pr = PrivacyRequest.get_by(
         db=db,
@@ -160,11 +159,11 @@ def test_intake_onetrust_requests_no_config(
 
 
 @mock.patch(
-    "fidesops.service.request_intake.onetrust_service.get_onetrust_access_token"
+    "fidesops.service.privacy_request.onetrust_service.get_onetrust_access_token"
 )
-@mock.patch("fidesops.service.request_intake.onetrust_service._get_all_requests")
-@mock.patch("fidesops.service.request_intake.onetrust_service._get_all_subtasks")
-@mock.patch("fidesops.models.privacy_request.PrivacyRequestRunner.run")
+@mock.patch("fidesops.service.privacy_request.onetrust_service.OneTrustService._get_all_requests")
+@mock.patch("fidesops.service.privacy_request.onetrust_service.OneTrustService._get_all_subtasks")
+@mock.patch("fidesops.service.privacy_request.request_runner_service.PrivacyRequestRunner.run")
 def test_intake_onetrust_requests_no_policy(
     finish_processing_mock: Mock,
     mock_get_all_subtasks: Mock,
@@ -199,7 +198,7 @@ def test_intake_onetrust_requests_no_policy(
     mock_get_all_subtasks.side_effect = [[mock_subtask_1], [mock_subtask_2]]
 
     with pytest.raises(PolicyNotFoundException):
-        intake_onetrust_requests(storage_config_onetrust.key)
+        OneTrustService.intake_onetrust_requests(storage_config_onetrust.key)
 
     pr = PrivacyRequest.get_by(
         db=db,
@@ -213,11 +212,11 @@ def test_intake_onetrust_requests_no_policy(
 
 
 @mock.patch(
-    "fidesops.service.request_intake.onetrust_service.get_onetrust_access_token"
+    "fidesops.service.privacy_request.onetrust_service.get_onetrust_access_token"
 )
-@mock.patch("fidesops.service.request_intake.onetrust_service._get_all_requests")
-@mock.patch("fidesops.service.request_intake.onetrust_service._get_all_subtasks")
-@mock.patch("fidesops.models.privacy_request.PrivacyRequestRunner.run")
+@mock.patch("fidesops.service.privacy_request.onetrust_service.OneTrustService._get_all_requests")
+@mock.patch("fidesops.service.privacy_request.onetrust_service.OneTrustService._get_all_subtasks")
+@mock.patch("fidesops.service.privacy_request.request_runner_service.PrivacyRequestRunner.run")
 def test_intake_onetrust_requests_auth_fail(
     finish_processing_mock: Mock,
     mock_get_all_subtasks: Mock,
@@ -253,7 +252,7 @@ def test_intake_onetrust_requests_auth_fail(
     mock_get_all_subtasks.side_effect = [[mock_subtask_1], [mock_subtask_2]]
 
     with pytest.raises(AuthenticationException):
-        intake_onetrust_requests(storage_config_onetrust.key)
+        OneTrustService.intake_onetrust_requests(storage_config_onetrust.key)
 
     pr = PrivacyRequest.get_by(
         db=db,
@@ -268,11 +267,11 @@ def test_intake_onetrust_requests_auth_fail(
 
 
 @mock.patch(
-    "fidesops.service.request_intake.onetrust_service.get_onetrust_access_token"
+    "fidesops.service.privacy_request.onetrust_service.get_onetrust_access_token"
 )
-@mock.patch("fidesops.service.request_intake.onetrust_service._get_all_requests")
-@mock.patch("fidesops.service.request_intake.onetrust_service._get_all_subtasks")
-@mock.patch("fidesops.models.privacy_request.PrivacyRequestRunner.run")
+@mock.patch("fidesops.service.privacy_request.onetrust_service.OneTrustService._get_all_requests")
+@mock.patch("fidesops.service.privacy_request.onetrust_service.OneTrustService._get_all_subtasks")
+@mock.patch("fidesops.service.privacy_request.request_runner_service.PrivacyRequestRunner.run")
 def test_intake_onetrust_requests_no_fides_tasks(
     finish_processing_mock: Mock,
     mock_get_all_subtasks: Mock,
@@ -309,7 +308,7 @@ def test_intake_onetrust_requests_no_fides_tasks(
     mock_get_all_subtasks.side_effect = [[mock_subtask_1], [mock_subtask_2]]
 
     # call test function
-    intake_onetrust_requests(storage_config_onetrust.key)
+    OneTrustService.intake_onetrust_requests(storage_config_onetrust.key)
 
     # assert
     mock_get_all_requests.assert_called_with(
